Move registry from the prototype architecture to the prototype's resources
