+++ panda/src/main/java/org/panda_lang/panda/framework/design/resource/prototypes/ClassPrototypeModel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.panda_lang.panda.framework.design.architecture.prototype.registry;
+package org.panda_lang.panda.framework.design.resource.prototypes;
 
 import org.panda_lang.panda.framework.PandaFramework;
 import org.panda_lang.panda.framework.design.architecture.prototype.method.MethodVisibility;
