+++ panda/src/main/java/org/panda_lang/panda/framework/language/resource/prototypes/ConsolePrototype.java
@@ -19,9 +19,9 @@ package org.panda_lang.panda.framework.language.resource.prototypes;
 import org.panda_lang.panda.framework.design.architecture.prototype.method.MethodVisibility;
 import org.panda_lang.panda.framework.design.architecture.value.Value;
 import org.panda_lang.panda.framework.design.runtime.ExecutableBranch;
-import org.panda_lang.panda.framework.design.architecture.prototype.registry.ClassPrototypeModel;
-import org.panda_lang.panda.framework.design.architecture.prototype.registry.ClassPrototypeModel.ClassDeclaration;
-import org.panda_lang.panda.framework.design.architecture.prototype.registry.ClassPrototypeModel.ModuleDeclaration;
+import org.panda_lang.panda.framework.design.resource.prototypes.ClassPrototypeModel;
+import org.panda_lang.panda.framework.design.resource.prototypes.ClassPrototypeModel.ClassDeclaration;
+import org.panda_lang.panda.framework.design.resource.prototypes.ClassPrototypeModel.ModuleDeclaration;
 
 import java.security.InvalidParameterException;
